requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
configparser
azure.servicebus 
requests
azure-keyvault
azure-identity
cx-Oracle==8.3.0